Skip quay source test on CI

+++ integration-tests/tests/create-application-from-quay-image.spec.ts
@@ -6,38 +6,41 @@ import { CreateApplicationPage } from '../support/pages/CreateApplicationPage';
 import { Applications } from '../utils/Applications';
 import { Common } from '../utils/Common';
 
-describe('Create Component from Quay Image', () => {
-  const addComponent = new AddComponentPage();
-  const componentPage = new ComponentPage();
-  const applicationDetailPage = new ApplicationDetailPage();
-  const applicationName = Common.generateAppName();
-  const quayImage = 'quay.io/quarkus/code-quarkus-app';
-  const componentName = 'code-quarkus-app';
-
-  before(function () {
-    const createApplicationPage = new CreateApplicationPage();
-    createApplicationPage.clickCreateApplication();
-    createApplicationPage.setApplicationName(applicationName);
-    createApplicationPage.clickNext();
-  });
-
-  after(function () {
-    Common.openApplicationURL(applicationName);
-    applicationDetailPage.deleteComponent(componentName);
-    applicationDetailPage.createdComponentNotExists(componentName);
-    Applications.deleteApplication(applicationName);
-  });
-
-  describe('Creating a Quay Component', () => {
-    it('Validate Repo', () => {
-      addComponent.setSource(quayImage);
-      addComponent.waitRepoValidated();
-      addComponent.clickNext();
+// skip on jenkins build due to SPI issues
+if (!Cypress.env('PR_CHECK')) {
+  describe('Create Component from Quay Image', () => {
+    const addComponent = new AddComponentPage();
+    const componentPage = new ComponentPage();
+    const applicationDetailPage = new ApplicationDetailPage();
+    const applicationName = Common.generateAppName();
+    const quayImage = 'quay.io/quarkus/code-quarkus-app';
+    const componentName = 'code-quarkus-app';
+  
+    before(function () {
+      const createApplicationPage = new CreateApplicationPage();
+      createApplicationPage.clickCreateApplication();
+      createApplicationPage.setApplicationName(applicationName);
+      createApplicationPage.clickNext();
     });
-
-    it('Create Application', () => {
-      componentPage.createApplication();
-      applicationDetailPage.createdComponentExists(componentName, applicationName);
+  
+    after(function () {
+      Common.openApplicationURL(applicationName);
+      applicationDetailPage.deleteComponent(componentName);
+      applicationDetailPage.createdComponentNotExists(componentName);
+      Applications.deleteApplication(applicationName);
+    });
+  
+    describe('Creating a Quay Component', () => {
+      it('Validate Repo', () => {
+        addComponent.setSource(quayImage);
+        addComponent.waitRepoValidated();
+        addComponent.clickNext();
+      });
+  
+      it('Create Application', () => {
+        componentPage.createApplication();
+        applicationDetailPage.createdComponentExists(componentName, applicationName);
+      });
     });
   });
-});
+}
